coords and dimensions now use floats

+++ src/main/java/team/lodestar/sage/client/gui/PositionInfo.java
@@ -1,11 +1,11 @@
 package team.lodestar.sage.client.gui;
 
 public class PositionInfo {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     public boolean relative; // relative to parent or not
-    public int paddingX;
-    public int paddingY;
-    public int width;
-    public int height;
+    public float paddingX;
+    public float paddingY;
+    public float width;
+    public float height;
 }
